requirements.txt
psycopg2-binary==2.9.11
psycopg2==2.9.11
faker==40.11.1
pandas==3.0.1
python-dotenv==1.2.2
sqlalchemy==2.0.48
openai==2.30.0
